+++ projects/ccd-case-ui-toolkit/src/lib/shared/pipes/pipes.module.ts
@@ -3,6 +3,8 @@ import { NgModule } from '@angular/core';
 import { CaseReferencePipe } from './case-reference';
 import { CcdCaseTitlePipe } from './case-title';
 import { CcdCollectionTableCaseFieldsFilterPipe, CcdCYAPageLabelFilterPipe, CcdPageFieldsPipe, CcdTabFieldsPipe, ReadFieldsFilterPipe } from './complex';
+import { LinkCasesFromReasonValuePipe } from './link-cases-from-reason-code/ccd-link-cases-from-reason-code.pipe';
+import { LinkCasesReasonValuePipe } from './link-cases-reason-code/ccd-link-cases-reason-code.pipe';
 import { SortSearchResultPipe } from './search-result/sorting/sort-search-result.pipe';
 
 const pipeDeclarations = [
@@ -13,7 +15,9 @@ const pipeDeclarations = [
   CcdCYAPageLabelFilterPipe,
   ReadFieldsFilterPipe,
   CcdTabFieldsPipe,
-  CcdPageFieldsPipe
+  CcdPageFieldsPipe,
+  LinkCasesReasonValuePipe,
+  LinkCasesFromReasonValuePipe
 ];
 
 @NgModule({